Example – Creating a soap web service bottoms up using endpoint publisher
http://www.javajee.com/articles/writing-a-simple-soap-web-service-bottoms-up

In this example (WS-Example-01) we will use bottom up approach to
develop a simple SOAP based web service and then publish it using an
Endpoint publisher. We will write a simple example web service for
getting the time from server. Here we will not specifying any SOAP
binding details and hence our web service will be a Document – Literal -
Wrapped web service, which is the default.