deploying a bottoms-up Web service in Tomcat (JEE container)
http://www.javajee.com/articles/deploying-a-bottoms-up-web-service-in-tomcat

The Gradle build for this project supports building a WAR file which can
be deployed to a Tomcat instance, but this WAR file may also be tested
with Gradle's Jetty test engine.

gradle jettyRun

A web container like tomcat can deploy servlets, jsps and web services.
Web components are generally deployed in the web container as war files.
These war files might contain standard configuration files such as
web.xml and vendor specific configuration files such as sun-jaxws.xml.
To host web services, a web container relies on a servlet interceptor,
like WSServlet for tomcat that mediates between the client and the web
service implementation class (SIB), which needs to be configured in the
web.xml file. However, with JAX-WS, this is taken care of "behind the
scenes" (stricken out in the example). The endpoint and url pattern
details are read from sun-jaxws.xml file and servlet configuration for
WSServlet has no effect here. The <endpoints> element of sun-jaxws.xml
contains one or more <endpoint> elements. Each endpoint represents a
port in the WSDL. The WSServletContextListener configured in the web.xml
as a listener class parses sun-jaxws.xml and sets up all deployed
endpoints.